requirements.txt
boto3==1.26.0
botocore==1.29.0
click==8.1.3
Flask==2.2.2
imageio==2.22.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
networkx==2.8.8
numpy==1.23.4
packaging==21.3
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
s3transfer==0.6.0
scikit-image==0.19.3
scipy==1.9.3
six==1.16.0
tifffile==2022.10.10
urllib3==1.26.12
Werkzeug==2.2.2
